removed frontend stuff, fixed pom, renamned cli files and splitted tests
