--- mycore-accesskey/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "mycore-accesskey",
-  "version": "1.0.0",
-  "description": "mycore-accesskey",
-  "devDependencies": {
-    "grunt": "^0.4.5",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-uglify-es": "^3.3.0"
-  },
-  "author": "Lennard Golsch <l.golsch@tu-braunschweig.de>",
-  "license": "GPL-3.0"
-}
