--- mycore-accesskey/src/main/resources/META-INF/resources/accesskey.xml
@@ -1,151 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE MyCoReWebPage>
-<MyCoReWebPage title="{i18n}">
-  <section i18n="mcr.accesskey.manage" xml:lang="all">
-    <div class="container-fluid">
-      <div class="row pb-2">
-        <div class="col-auto mr-auto">
-          <h2>
-            <i18n key="mcr.accesskey.manage" />
-          </h2>
-        </div>
-        <div class="col-auto">
-          <button type="button" class="btn btn-primary float-right" id="new-btn" href="#key-modal" data-toggle="modal" data-mode="new" style="display:none">
-            <i class="fa fa-plus"></i>
-          </button>
-        </div>
-      </div>
-      <div id="alert-div" class="alert alert-warning" style="display: none;">
-        <h1><i18n key="mcr.accesskey.headline.401" /></h1>
-        <p><i18n key="mcr.accesskey.error.main" /></p>
-      </div>
-      <table id="key-table" class="table table-striped" style="display:none">
-        <thead>
-          <tr class="d-flex">
-            <th class="th col-auto">#</th>
-            <th class="th col-2 text-center">
-              <i18n key="mcr.accesskey.type" />
-            </th>
-            <th class="th col text-center">
-              <i18n key="mcr.accesskey.value" />
-            </th>
-            <th class="th col-auto" />
-          </tr>
-        </thead>
-        <tbody>
-          <tr class="d-flex">
-            <td class="th-sm col-auto"></td>
-            <td class="th-sm col-2 text-center"></td>
-            <td class="th-sm col text-center"></td>
-            <td class="th-sm col-auto">
-              <a class="edit-btn" href="#key-modal" data-toggle="modal">
-                <i class="fas fa-edit"></i>
-              </a>
-            </td>
-          </tr>
-        </tbody>
-      </table>
-      <div class="d-flex justify-content-center">
-        <div id="spinner" class="spinner-border" role="status" style="margin-top:200px">
-          <span class="sr-only">Loading...</span>
-        </div>
-      </div>
-      <div id="paginator-div" style="display:none">
-        <hr class="my-4" />
-        <nav aria-label="Page navigation example">
-          <ul class="pagination justify-content-center">
-            <li class="page-item" id="pre">
-              <a class="page-link" href="#" tabindex="-1">
-                <i class="fas fa-chevron-left"></i>
-              </a>
-            </li>
-            <li class="page-item"><a class="page-link" href="#">1</a></li>
-            <li class="page-item"><a class="page-link" href="#">2</a></li>
-            <li class="page-item"><a class="page-link" href="#">3</a></li>
-            <li class="page-item"><a class="page-link" href="#">4</a></li>
-            <li class="page-item"><a class="page-link" href="#">5</a></li>
-            <li class="page-item"><a class="page-link" href="#">6</a></li>
-            <li class="page-item"><a class="page-link" href="#">7</a></li>
-            <li class="page-item"><a class="page-link" href="#">8</a></li>
-            <li class="page-item"><a class="page-link" href="#">9</a></li>
-            <li class="page-item" id="next">
-              <a class="page-link" href="#">
-                <i class="fas fa-chevron-right"></i>
-              </a>
-            </li>
-          </ul>
-        </nav>
-      </div>
-    </div>
-
-    <!-- Modal -->
-    <div class="modal fade" id="key-modal" tabindex="-1" aria-labelledby="key-modal" aria-hidden="true"> 
-      <div class="modal-dialog modal-lg">
-        <div class="modal-content">
-          <div class="modal-header">
-            <h5 class="modal-title">
-              <i18n key="mcr.accesskey.edit" />
-            </h5>
-            <button type="button" class="close modal-close-btn" data-dismiss="modal" aria-label="Close">
-              <span aria-hidden="true">X</span>
-            </button>
-          </div>
-          <div class="modal-body">
-            <div class="container-fluid">
-              <div id="modal-alert-div" class="alert alert-danger" style="display: none;" role="alert">
-                This is a danger alert—check it out!
-              </div>
-              <input type="hidden" id="id-input" />
-              <div class="form-row">
-                <div class="col-4 form-group">
-                  <label for="type-input"><i18n key="mcr.accesskey.type" />*</label>
-                  <select id="type-input" class="form-control">
-                    <option value="read">read</option>
-                    <option value="writedb">write</option>
-                  </select>
-                </div>
-              </div>
-              <div class="row">
-                <div class="col form-group">
-                  <label for="value-input"><i18n key="mcr.accesskey.value" />*</label>
-                  <div class="input-group">
-                    <span class="input-group-prepend">
-                      <button aria-label="Generate" class="btn btn-secondary" id="value-gen-btn" title="Generate" type="button">
-                        <span aria-hidden="true" class="fas fa-cogs"></span>
-                        <span class="sr-only">
-                          <i18n key="mcr.accesskey.generate" />
-                        </span>
-                      </button>
-                    </span>
-                    <input id="value-input" type="text" class="form-control" />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="modal-footer">
-            <div class="container-fluid">
-              <div class="row">
-                <div class="col">
-                  <button id="delete-btn" type="button" class="btn btn-danger">
-                    <i18n key="mcr.accesskey.delete" />
-                  </button>
-                </div>
-                <div class="col-auto">
-                  <button id="update-btn" type="button" class="btn btn-primary">
-                    Update
-                  </button>
-                  <button id="add-btn" type="button" class="btn btn-primary">
-                    <i18n key="mcr.accesskey.add" />
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <script type="text/javascript" src="accesskey-utils.min.js" />
-    <script type="text/javascript" src="accesskey.min.js" />
-  </section>
-</MyCoReWebPage>
