--- mycore-accesskey/src/main/resources/META-INF/web-fragment.xml
@@ -1,12 +1,0 @@
-<web-fragment xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-fragment_3_0.xsd" version="3.0" metadata-complete="true"
->
-  <servlet>
-    <servlet-name>MCRAccessKeyServlet</servlet-name>
-    <servlet-class>org.mycore.accesskey.frontend.servlet.MCRAccessKeyServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>MCRAccessKeyServlet</servlet-name>
-    <url-pattern>/servlets/MCRAccessKeyServlet</url-pattern>
-  </servlet-mapping>
-</web-fragment>
