+++ mycore-accesskey/src/main/resources/xsl/save-accesskey-object.xsl
