--- mycore-accesskey/src/test/resources/mycore.properties
@@ -1,1 +1,0 @@
-MCR.Access.Strategy.Class=org.mycore.accesskey.strategy.MCRAccessKeyStrategy
